Move to type_ref
