+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/mod.rs
@@ -7,7 +7,6 @@ mod get_span;
 mod to_candid_type;
 mod ts_fn_or_constructor_type;
 mod ts_type_lit;
-mod ts_type_ref;
 mod type_to_string;
 
 impl<'a> SourceMapped<'a, TsType> {
