--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/ts_type_ref/get_name.rs
@@ -1,20 +1,0 @@
-use swc_ecma_ast::{TsEntityName, TsTypeRef};
-
-use crate::{traits::GetName, ts_ast::SourceMapped};
-
-impl GetName for SourceMapped<'_, TsTypeRef> {
-    fn get_name(&self) -> &str {
-        match &self.type_name {
-            TsEntityName::TsQualifiedName(ts_qualified_name) => {
-                // TODO: This could be improved for Qualified TypeRefs with type params.
-                // Currently we just drop the type params. It would be better if we
-                // included them.
-                panic!(
-                    "{}",
-                    self.qualified_name_error(ts_qualified_name.right.get_name().to_string())
-                )
-            }
-            TsEntityName::Ident(identifier) => identifier.get_name(),
-        }
-    }
-}
