--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/ts_type_ref/get_span.rs
@@ -1,10 +1,0 @@
-use swc_common::Span;
-use swc_ecma_ast::TsTypeRef;
-
-use crate::traits::GetSpan;
-
-impl GetSpan for TsTypeRef {
-    fn get_span(&self) -> Span {
-        self.span
-    }
-}
